Move playwright fixtures in to `mavis.test`
This moves the fixtures related to setting up Playwright in to their own
file.
