+++ conftest.py
@@ -1,47 +1,4 @@
-import os
-from pathlib import Path
-from typing import Optional
-
 import pytest
-
-from mavis.test.playwright_ops import PlaywrightOperations
-from mavis.test.wrappers import get_current_datetime
-
-
-@pytest.fixture(scope="session")
-def base_url() -> str:
-    return os.environ["BASE_URL"]
-
-
-@pytest.fixture(scope="session")
-def basic_auth() -> dict[str, str]:
-    return {
-        "username": os.environ["BASIC_AUTH_USERNAME"],
-        "password": os.environ["BASIC_AUTH_PASSWORD"],
-    }
-
-
-@pytest.fixture(scope="session")
-def screenshots_path(pytestconfig, browser_name: str) -> Optional[Path]:
-    screenshot = pytestconfig.getoption("screenshot")
-
-    if screenshot == "off":
-        return None
-
-    session_name = f"{get_current_datetime()}-{browser_name}"
-    path = Path("screenshots") / session_name
-    path.mkdir(parents=True, exist_ok=True)
-    return path
-
-
-@pytest.fixture(scope="session")
-def browser_context_args(browser_context_args, basic_auth):
-    return {**browser_context_args, "http_credentials": basic_auth}
-
-
-@pytest.fixture
-def playwright_operations(page, screenshots_path):
-    return PlaywrightOperations(page, screenshots_path)
 
 
 @pytest.fixture
