+++ mavis/test/fixtures/__init__.py
@@ -1,0 +1,15 @@
+from .playwright import (
+    base_url,
+    basic_auth,
+    browser_context_args,
+    playwright_operations,
+    screenshots_path,
+)
+
+__all__ = [
+    "base_url",
+    "basic_auth",
+    "browser_context_args",
+    "playwright_operations",
+    "screenshots_path",
+]
